+++ arm-sysregs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 - Added the PAN register
+- Added `is_*_present` for `FEAT_PAN`, `FEAT_PAN2` and `FEAT_PAN3`.
 
 ## 0.4.0